requirements.txt
fastapi
uvicorn
python-dotenv
twilio
boto3
<<<<<<< HEAD
=======
SQLAlchemy
psycopg2-binary
>>>>>>> 1acc33f (Initial commit - FastAPI TVS Mitra VoiceBot)
